Move the extrude_polygon() function from AdhocShape to Solid, remove AdhocShape completely
