+++ crates/opencascade/src/lib.rs
@@ -1,6 +1,5 @@
 use thiserror::Error;
 
-pub mod adhoc;
 pub mod angle;
 pub mod mesh;
 pub mod primitives;
